dcolocate states and functions closer to components by adding separate context
