+++ client/src/components/ChatWindow.js
@@ -1,41 +1,11 @@
-import { useEffect } from 'react';
 import FilterButton from './FilterButton';
 import Form from './Form';
 import OpenChat from './OpenChat';
-import { useMessaging } from '../hooks/useMessaging';
-import { useGlobalContext } from '../context/OverallContext';
+import { useMessageContext } from '../context/MessageContext';
 
 const ChatWindow = () => {
-  const {
-    API_MESSAGES,
-    userId,
-    jwt,
-    chat,
-    newMessage,
-    setNewMessage,
-    setIsMessageSent,
-  } = useGlobalContext();
-  const { postMessage, fetchMessages } = useMessaging();
-
-  const handleMessage = (e) => {
-    setNewMessage({ ...newMessage, [e.target.name]: e.target.value });
-  };
-
-  const sendMessage = (e) => {
-    e.preventDefault();
-    postMessage(API_MESSAGES, chat._id, jwt, newMessage);
-    setIsMessageSent(true);
-  };
-
-  const handleKeyPress = (e) => {
-    if (e.charCode === 13) {
-      sendMessage(e);
-    }
-  };
-
-  useEffect(() => {
-    fetchMessages(API_MESSAGES, sessionStorage.getItem('convId'), jwt, userId);
-  }, [API_MESSAGES, fetchMessages, jwt, userId]);
+  const { sendMessage, newMessage, handleKeyPress, handleMessage } =
+    useMessageContext();
 
   return (
     <>
