+++ client/src/components/GenreFilter.js
@@ -1,22 +1,8 @@
-import { useState } from 'react';
 import FilterButton from './FilterButton';
-import { useGlobalContext } from '../context/OverallContext';
+import { useMarketplaceContext } from '../context/MarketplaceContext';
 
 const GenreFilter = () => {
-  const { allBooks, setBooks } = useGlobalContext();
-  const allCategories = [
-    'alle',
-    ...new Set(allBooks.map((book) => book.category)),
-  ];
-  const [categories] = useState(allCategories);
-
-  const filterBooks = (category) => {
-    if (category === 'alle') {
-      return setBooks(allBooks);
-    }
-    let filteredBooks = allBooks.filter((book) => book.category === category);
-    setBooks(filteredBooks);
-  };
+  const { categories, filterBooks } = useMarketplaceContext();
 
   return (
     <>
