+++ client/src/components/MenuLink.js
@@ -1,5 +1,5 @@
 import { NavLink } from 'react-router-dom';
-import { useGlobalContext } from '../context/OverallContext';
+import { useGlobalContext } from '../context/GlobalContext';
 
 const MenuLink = ({ id, url, text }) => {
   const { hideLinks } = useGlobalContext();
