+++ client/src/components/UserAction.js
@@ -1,39 +1,19 @@
-import { useEffect } from 'react';
-import { useGlobalContext } from '../context/OverallContext';
-import { useBookData } from '../hooks/useBookData';
+import { useGlobalContext } from '../context/GlobalContext';
+import { useOpenBookContext } from '../context/OpenBookContext';
 import ActionButton from './ActionBtn';
 
-const UserAction = ({ id, owner, condition }) => {
-  const {
-    userId,
-    user,
-    jwt,
-    API_BOOKS,
-    API_USERS,
-    setShowMessageModal,
-    setShowEditBook,
-  } = useGlobalContext();
-  const { fetchUser, deleteSingleBook } = useBookData();
-
-  const removeBook = () => {
-    deleteSingleBook(API_BOOKS, id, jwt);
-  };
-
-  const messageUser = () => {
-    sessionStorage.setItem('receiver', owner);
-    setShowMessageModal(true);
-  };
-
-  useEffect(() => {
-    fetchUser(owner, API_USERS, jwt);
-  }, [owner, API_USERS, jwt, fetchUser]);
+const UserAction = () => {
+  const { userId } = useGlobalContext();
+  const { openBook, removeBook, openEditWindow, messageUser } =
+    useOpenBookContext();
+  const { username, condition, owner } = openBook;
 
   return (
     <>
       <aside className='user-action'>
         <section className='action-section'>
           <p>Dieses Buch gehört:</p>
-          <h3>{user}</h3>
+          <h3>{username}</h3>
         </section>
         <hr className='separation-line' />
         <section className='action-section'>
@@ -45,11 +25,7 @@ const UserAction = ({ id, owner, condition }) => {
           <p>Was möchtest du tun?</p>
           {owner === userId ? (
             <>
-              <ActionButton
-                onClick={() => {
-                  setShowEditBook(true);
-                }}
-              >
+              <ActionButton onClick={openEditWindow}>
                 Jetzt bearbeiten
               </ActionButton>
               <ActionButton onClick={removeBook}>Jetzt löschen</ActionButton>
