+++ client/src/components/UserDashboard.js
@@ -1,4 +1,4 @@
-import { useGlobalContext } from '../context/OverallContext';
+import { useGlobalContext } from '../context/GlobalContext';
 
 const UserDashboard = () => {
   const { userName } = useGlobalContext();
