+++ client/src/utils/linksDB.js
@@ -4,9 +4,8 @@ const links = [
   { id: 3, url: '/messages', text: 'nachrichten' },
 ];
 const footer = [
-  { id: 1, url: '/about', text: 'über uns' },
-  { id: 2, url: '/imprint', text: 'impressum' },
-  { id: 3, url: '/dataprivacy', text: 'datenschutz' },
+  { id: 1, url: '/imprint', text: 'impressum' },
+  { id: 2, url: '/dataprivacy', text: 'datenschutz' },
 ];
 
 export { links, footer };
