+++ server/models/conversation.model.js
@@ -1,22 +1,24 @@
-import mongoose from 'mongoose'
-import User from '../models/user.model'
-
+import mongoose from 'mongoose';
 
 const ConversationSchema = new mongoose.Schema({
-    recipients: [{
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User',
-    }],
-    messages: [{
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'Message'
-    }],
-    created: {
-        type: Date,
-        default: Date.now,
+  recipients: [
+    {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User',
     },
-    // BookId als betreff
-    updated: Date // Last message
-})
+  ],
+  messages: [
+    {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'Message',
+    },
+  ],
+  created: {
+    type: Date,
+    default: Date.now,
+  },
+  // BookId als betreff
+  updated: Date, // Last message
+});
 
-export default mongoose.model('Conversation', ConversationSchema)
+export default mongoose.model('Conversation', ConversationSchema);
