+++ server/models/messages.model.js
@@ -1,23 +1,22 @@
-import mongoose from 'mongoose'
-import User from '../models/user.model'
+import mongoose from 'mongoose';
 
 const MessageSchema = new mongoose.Schema({
-    sender: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User',
-    },
-    reciever: {
-        type: mongoose.Schema.Types.ObjectId,
-        ref: 'User',
-    },
-    message: {
-        type: String,
-        required: 'Bitte Nachricht eingeben',
-    },
-    created: {
-        type: Date,
-        default: Date.now,
-    }
-})
+  sender: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+  },
+  reciever: {
+    type: mongoose.Schema.Types.ObjectId,
+    ref: 'User',
+  },
+  message: {
+    type: String,
+    required: 'Bitte Nachricht eingeben',
+  },
+  created: {
+    type: Date,
+    default: Date.now,
+  },
+});
 
-export default mongoose.model('Message', MessageSchema)
+export default mongoose.model('Message', MessageSchema);
